Change default port to 9684
The port number was marked in the Prometheus Github documentation
as the one for this exporter.

Additionally, a Makefile was added to make build and release process
easier. The version of the binary was set to

  <kibana_minor_version>.x.<exporter_version>

For an example, currently the version would be 7.5.x.1 since the API
that this is tested is 7.5 and the version of the binary is 1.