Add flag to block until Kibana comes up
This flag was to address the need to fail early, however it evolved into
a blocking wait since a fail early wouldn't provide much of a benefit in
a CMS where a state machine would just start another Container if there
is only a failure, resuling in a crash loopback.

The flag is set to false by default, however I'm not sure if this should
be enabled by default. The more I think of it, enabling it by default
would be more beneficial. Maybe in the next major release.