Refactor collector code to be more readable
This change adds code quality in terms of decoupled logic and testable
pieces.

The unit tests were slightly improved to be based on a data set rather
than a series of repeated code.

The integration tests are yet to be added, since a Dockerized approach
would probably be needed instead of the current manual one.