Add GH action to run tests on push
This is will be the starting point to use a fully automated workflow
starting from tagging. Idea is to minimise the work I do on the
workstation and actually follow my own ramblings on devops.

Also tested the exporter against Kibana/ES 8.5.3 and bumped the version
in the Makefile. This will be replaced with a git ref lookup in the
future.